+++ Sources/MapboxCoreNavigation/NativeHandlersFactory.swift
@@ -1,3 +1,4 @@
+import MapboxCommon
 import MapboxNavigationNative
 import MapboxDirections
 import Foundation
@@ -13,14 +14,14 @@ class NativeHandlersFactory {
     // MARK: - Settings
     
     let tileStorePath: String
-    let credentials: DirectionsCredentials
+    let credentials: Credentials
     let tilesVersion: String?
     let historyDirectoryURL: URL?
     let targetVersion: String?
     let configFactoryType: ConfigFactory.Type
     
     init(tileStorePath: String,
-         credentials: DirectionsCredentials,
+         credentials: Credentials,
          tilesVersion: String? = nil,
          historyDirectoryURL: URL? = nil,
          targetVersion: String? = nil,