+++ Sources/MapboxCoreNavigation/RouteController.swift
@@ -3,6 +3,7 @@
 
 import Foundation
 import CoreLocation
+import MapboxCommon
 import MapboxNavigationNative
 import MapboxMobileEvents
 import MapboxDirections
@@ -503,7 +504,7 @@ open class RouteController: NSObject {
         }
     }
     
-    func mode(_ profileIdentifier: DirectionsProfileIdentifier) -> ActiveGuidanceMode {
+    func mode(_ profileIdentifier: ProfileIdentifier) -> ActiveGuidanceMode {
         switch profileIdentifier {
         case .automobile:
             return .driving