+++ Sources/MapboxCoreNavigation/TileStore.swift
@@ -1,3 +1,4 @@
+import MapboxCommon
 import MapboxNavigationNative